ml/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["syllable_counter"]

[project]
name = "syllable-counter"
version = "0.1.0"
description = "ML-based syllable counter for Todo Haiku"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Todo Haiku Team" },
]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.2.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "onnx-weekly>=1.19.0.dev20250425",
    "onnxruntime>=1.20.1",
    "rich>=14.0.0",
    "nltk>=3.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "ruff>=0.0.270",
]

# Full dependencies including ONNX
full = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "ruff>=0.0.270",
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
